+++ changelog.md
@@ -6,6 +6,7 @@
 - Fixed Balloon shaders (Malte)
 - Fixed Steel Axe having overly high damage (BluSunrize)
 - Fixed potential crashes related to the new blueprint renders (BluSunrize)
+- Fixed coresample markers disappearing off maps after restarting (BluSunrize)
 
 #####Version 0.12-88 - BUILT
 - Changed Multiblocks to accept redstone-ignoring conveyors as well as default ones. They will revert to normal ones when the multiblock is broken. (BluSunrize)